fix: relocated StringCollectionUtilTest in JCF: Collection
